Rearrange src folder, add react-app babel preset, change render prop for children
