+++ src/context/index.js
@@ -1,0 +1,1 @@
+export default ({ reducer, actions, saga, middleware, devToolsOptions }) => {}
